+++ .github/docs-site/javascripts/repository-facts.js
@@ -1,7 +1,7 @@
 (() => {
   "use strict";
 
-  const RELEASE_VERSION = "v7.0.0";
+  const RELEASE_VERSION = "v7.1.0";
   const SOURCE_CACHE_MARKER = "__source";
   const VERSION_SELECTOR = ".md-source__fact--version";